requirements.txt
Flask
Flask-SQLAlchemy
requests
pdfkit
python-dotenv
Flask-WTF
Werkzeug
